Refactor publishing api client
We don't need this class to be abstracted into an ItemService class and
can use it directly from the preload_items_processor. We also move the
publishing api client into the ETL app structure to be aligned with
changes in #654
