+++ app/services/items_service.rb
@@ -1,13 +1,7 @@
 require 'gds_api/content_store'
 class ItemsService
   def initialize
-    @publishing_api_client = Clients::PublishingAPI.new
     @content_store_client = GdsApi::ContentStore.new(Plek.new.find('content-store'))
-  end
-
-  def fetch_all
-    publishing_api_client
-      .fetch_all(%w[content_id base_path locale])
   end
 
   def fetch_raw_json(base_path)
@@ -16,5 +10,5 @@ class ItemsService
 
 private
 
-  attr_reader :publishing_api_client, :content_store_client
+  attr_reader :content_store_client
 end
